sec_kg_chat/requirements.txt
langchain-community>=0.0.10
langchain-openai>=0.0.2
python-dotenv>=1.0.0
PyYAML>=6.0
neo4j>=5.0.0
openai>=1.0.0